docker-container/FbClaw/requirements.txt
agno>=1.3.0
mistralai>=1.0.0
groq>=0.11.0
python-telegram-bot>=21.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
